+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialWorldSettings.h
@@ -6,6 +6,7 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/WorldSettings.h"
+#include "Utils/LayerInfo.h"
 
 #include "SpatialWorldSettings.generated.h"
 
@@ -18,13 +19,17 @@ class SPATIALGDK_API ASpatialWorldSettings : public AWorldSettings
 	GENERATED_BODY()
 
 public:
-	UPROPERTY(EditAnywhere, Config, Category = "Load Balancing")
-	TSubclassOf<UAbstractLBStrategy> LoadBalanceStrategy;
+	/** Enable running different server worker types to split the simulation. */
+	UPROPERTY(EditAnywhere, Config, Category = "Multi-Worker")
+	bool bEnableMultiWorker;
 
-	UPROPERTY(EditAnywhere, Config, Category = "Load Balancing")
-	TSubclassOf<UAbstractLockingPolicy> LockingPolicy;
+	UPROPERTY(EditAnywhere, Config, Category = "Multi-Worker", meta = (EditCondition = "bEnableMultiWorker"))
+	TSubclassOf<UAbstractLBStrategy> DefaultLayerLoadBalanceStrategy;
 
-	/** Layer load balancing  configuration. */
-	UPROPERTY(EditAnywhere, Config, Category = "Multi Worker")
-	TMap<FName, FLBLayerInfo> WorkerLBLayers;
+	UPROPERTY(EditAnywhere, Config, Category = "Multi-Worker", meta = (EditCondition = "bEnableMultiWorker"))
+	TSubclassOf<UAbstractLockingPolicy> DefaultLayerLockingPolicy;
+
+	/** Layer configuration. */
+	UPROPERTY(EditAnywhere, Config, Category = "Multi-Worker", meta = (EditCondition = "bEnableMultiWorker"))
+	TMap<FName, FLayerInfo> WorkerLayers;
 };