+++ SpatialGDK/Source/SpatialGDK/Public/Interop/SpatialSender.h
@@ -27,7 +27,6 @@ class USpatialPackageMapClient;
 class USpatialReceiver;
 class USpatialStaticComponentView;
 class USpatialClassInfoManager;
-class SpatialActorGroupManager;
 class USpatialWorkerConnection;
 
 struct FReliableRPCForRetry
@@ -178,8 +177,6 @@ private:
 
 	UPROPERTY()
 	USpatialClassInfoManager* ClassInfoManager;
-
-	SpatialActorGroupManager* ActorGroupManager;
 
 	FTimerManager* TimerManager;